+++ firmware/colibri/src/utils/crypto.cpp
@@ -62,7 +62,7 @@ std::string generateMnemonic(uint8_t words) {
 
   uint8_t entropy[32] = {0};
   generateEntropy(entropy);
-  // log_s("Entropy: %s", toHex(entropy, sizeof(entropy)).c_str());
+  log_ss("Entropy: %s", toHex(entropy, sizeof(entropy)).c_str());
 
   const char *mnemonic = mnemonic_from_data(entropy, strength / 8);
 
@@ -70,7 +70,7 @@ std::string generateMnemonic(uint8_t words) {
   memzero(entropy, sizeof(entropy));
   mnemonic_clear();
 
-  // log_s("Generated mnemonic phrase: %s", m.c_str());
+  log_ss("Generated mnemonic phrase: %s", m.c_str());
 
   return m;
 }
@@ -130,4 +130,12 @@ size_t fromHex(const char *hex, uint8_t *bytes, size_t maxBytesLen) {
   }
 
   return bytesLen;
+}
+
+// Helper method to convert uint32_t to bytes
+void uint32ToBytes(uint32_t value, uint8_t output[4]) {
+  output[0] = (value >> 24) & 0xFF;
+  output[1] = (value >> 16) & 0xFF;
+  output[2] = (value >> 8) & 0xFF;
+  output[3] = value & 0xFF;
 }