+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 # Colibri Changelog
 
-## 0.0.1
+(`*` = new feature; `+` = improvement; `#` = bugfix; `~` = chore)
+
+## v0.0.2
+
+2024-12-19
+
+- `*` Interfaces: finished communication via BLE
+- `*` Wallet: added optional self-destruct mode after n failed login attempts
+- `*` Wallet: added option to overwrite existing mnemonics
+- `*` Wallet: added possibility to change password and re-encrypt existing mnemonics
+- `+` Debug: added "super sensitive" log category
+- `+` RPC: various improvements for methods, descriptions, params and return values
+- `+` Storage: added support for custom NVS partition
+- `~` Base Typescript monorepo setup for SDK- and Webapp-development
+
+## v0.0.1
 
 2024-12-16